requirements.txt
# Web scraping
selenium==4.17.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2

# API y servidor web
flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
gunicorn==21.2.0
pyjwt==2.8.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0